+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 * Store :attr:`Env.session_info` when it is retrieved with Webclient API.
   Insert ``user_context``  into ``session_info`` when RPC API is used.
 
+* Method ``exists`` becomes private with Odoo 19.0.
+
 
 2.3.2 (2025-10-01)
 ~~~~~~~~~~~~~~~~~~